Improved prediction accuracy module handling by creating a factory, runner, noop, and default impl
