+++ libs/core/src/main/java/org/transitclock/core/prediction/accuracy/PredAccuracyPrediction.java
@@ -1,6 +1,8 @@
 /* (C)2023 */
 package org.transitclock.core.prediction.accuracy;
 
+import lombok.Value;
+
 import java.util.Date;
 
 /**
