+++ libs/core/src/main/java/org/transitclock/core/prediction/accuracy/PredictionAccuracyModuleRunnable.java
@@ -1,0 +1,5 @@
+package org.transitclock.core.prediction.accuracy;
+
+interface PredictionAccuracyModuleRunnable {
+    void run();
+}
