+++ libs/core/src/test/resources/testConfig.xml
@@ -5,7 +5,7 @@
         <!-- GtfsRealtimeModule read a GTFS vehicle location feed as the source of AVL data. -->
         <!-- PredictionAccuracyModule is used to store the transiTime prediction and record their accuracy. Look in class for comments explaining detail. -->
         <!-- GTFSRealtimePredictionAccuracyModule is used to store an alternative source of predictions and again record there accuracy based on actual arrival/departure times -->
-        <optionalModulesList>org.transitclock.avl.GtfsRealtimeModule;org.transitclock.core.prediction.accuracy.gtfsrt.GTFSRealtimePredictionAccuracyModule</optionalModulesList>
+        <optionalModulesList>org.transitclock.avl.GtfsRealtimeModule;org.transitclock.core.prediction.accuracy.GTFSRealtimePredictionAccuracyModule</optionalModulesList>
     </modules>
     <predAccuracy>
     	<pollingRateMsec>30000</pollingRateMsec>
